Second-pass Python 2 audit: fix 3 more live crash bugs, remove dead code
Broader sweep this time (shinken/ + modules/, not just the daemons
touched so far):

- shinken/external_command.py: 5 calls to long(value) in
  CHANGE_*_MODATTR handlers -- long doesn't exist in Python 3, so any
  CHANGE_SVC_MODATTR/CHANGE_HOST_MODATTR/CHANGE_CONTACT_MODATTR external
  command would crash with NameError. -> int(value).
- shinken/discovery/discoverymanager.py: get_uuid() fallback used
  sys.maxint (renamed sys.maxsize in Python 3) -> AttributeError if the
  uuid module were ever unavailable.
- modules/livestatus/module.py: exception handler referenced
  os.errno.EINTR; os.errno doesn't exist in current Python 3 (confirmed
  by testing) -> the error-handling path for an interrupted queue read
  would itself crash instead of handling the original error. errno was
  already imported in the file, just unused for this.

Also removed shinken/sorteddict.py entirely: a full SortedDict class
with a Python-2-style API (.iterkeys/.itervalues/.iteritems as its own
methods, harmless in isolation) that turned out to be imported by
nothing anywhere in the repo. Dead code, not a bug, but no reason to
carry a whole unused file across the port.

Confirmed via full rebuild + restart that the stack stays stable, and
specifically exercised the previously-broken CHANGE_SVC_MODATTR command
to confirm the int(value) fix actually works end-to-end.

Left as-is (not bugs, just historical debt): shinken/check.py has a
commented-out `# __metaclass__ = AutoSlots` with an existing
`# FIXME : reenable AutoSlots if possible` -- already disabled and
flagged by a prior author, re-enabling it is a separate, larger effort
than this audit's scope.

+++ shinken/external_command.py
@@ -784,15 +784,15 @@ class ExternalCommandManager:
 
     # CHANGE_CONTACT_MODSATTR;<contact_name>;<value>
     def CHANGE_CONTACT_MODSATTR(self, contact, value):  # TODO
-        contact.modified_service_attributes = long(value)
+        contact.modified_service_attributes = int(value)
 
     # CHANGE_CONTACT_MODHATTR;<contact_name>;<value>
     def CHANGE_CONTACT_MODHATTR(self, contact, value):  # TODO
-        contact.modified_host_attributes = long(value)
+        contact.modified_host_attributes = int(value)
 
     # CHANGE_CONTACT_MODATTR;<contact_name>;<value>
     def CHANGE_CONTACT_MODATTR(self, contact, value):
-        contact.modified_attributes = long(value)
+        contact.modified_attributes = int(value)
 
     # CHANGE_CONTACT_HOST_NOTIFICATION_TIMEPERIOD;<contact_name>;<notification_timeperiod>
     def CHANGE_CONTACT_HOST_NOTIFICATION_TIMEPERIOD(self, contact, notification_timeperiod):
@@ -888,7 +888,7 @@ class ExternalCommandManager:
 
     # CHANGE_HOST_MODATTR;<host_name>;<value>
     def CHANGE_HOST_MODATTR(self, host, value):
-        host.modified_attributes = long(value)
+        host.modified_attributes = int(value)
 
     # CHANGE_MAX_HOST_CHECK_ATTEMPTS;<host_name>;<check_attempts>
     def CHANGE_MAX_HOST_CHECK_ATTEMPTS(self, host, check_attempts):
@@ -964,7 +964,7 @@ class ExternalCommandManager:
         # This is not enough.
         # We need to also change each of the needed attributes.
         previous_value = service.modified_attributes
-        future_value = long(value)
+        future_value = int(value)
         changes = future_value ^ previous_value
 
         for modattr in [
